backend/get-talent/requirements_cvex.txt
Django==4.1.4
djangorestframework==3.13.1
psycopg2-binary==2.9.3
PyMuPDF==1.23.6
PyPDF2==2.11.0
selenium==4.2.0
docxtpl==0.16.4
django-cors-headers==3.13.0
allcities==1.0.3
gunicorn==20.1.0
pdftotext==2.2.2
openpyxl==3.1.2
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
blis==0.7.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.5
confection==0.1.0
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
cytoolz==0.12.1
decorator==5.1.1
favicon==0.7.0
floret==0.10.3
fonttools==4.41.0
gitdb==4.0.10
GitPython==3.1.32
htbuilder==0.6.1
idna==3.4
importlib-metadata==6.8.0
jellyfish==1.0.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
langcodes==3.3.0
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==9.1.0
murmurhash==1.0.9
networkx==3.1
packaging==23.1
pandas==2.0.3
pathy==0.10.2
Pillow==10.0.0
plotly==5.15.0
preshed==3.0.8
protobuf==4.23.4
pyarrow==12.0.1
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.15.1
pymdown-extensions==10.1
Pympler==1.0.1
pyparsing==3.0.9
pypdf==3.12.2
pyphen==0.14.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rpds-py==0.8.11
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
soupsieve==2.4.1
srsly==2.4.6
st-annotated-text==4.0.0
tenacity==8.2.2
thinc==8.1.10
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3>=1.26.14
validators==0.20.0
wasabi==1.1.2
watchdog==3.0.0
zipp==3.16.2
reportlab==4.0.4
easygui==0.98.3
pyocclient==0.6
typing-extensions==4.7.1 #####
python-dotenv==1.0.0 #####
djangorestframework-simplejwt==5.3.1
fpdf==1.7.2
doc2pdf==0.2.0
pgvector==0.2.5
pdf2image==1.17.0
paddleocr==2.7.3
paddlepaddle==2.6.1
apify-client==1.7.0
whitenoise==6.7.0